Rename file in `state_processing`
