+++ eth2/state_processing/src/common/initiate_validator_exit.rs
