+++ eth2/state_processing/src/common/mod.rs
@@ -1,11 +1,11 @@
 mod convert_to_indexed;
-mod exit;
 mod get_attesting_indices;
+mod initiate_validator_exit;
 mod slash_validator;
 mod verify_bitfield;
 
 pub use convert_to_indexed::convert_to_indexed;
-pub use exit::initiate_validator_exit;
 pub use get_attesting_indices::{get_attesting_indices, get_attesting_indices_unsorted};
+pub use initiate_validator_exit::initiate_validator_exit;
 pub use slash_validator::slash_validator;
 pub use verify_bitfield::verify_bitfield_length;
